SLEEP DROUSING PROJECT/requirements.txt
# Driver Drowsiness Detection System — Python Dependencies
# Install all at once: pip install -r requirements.txt

opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
pyserial>=3.5
